requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
requests>=2.28.1
gradio>=3.0.0
networkx>=2.8.8
pyvis>=0.3.2
redis>=4.5.4
matplotlib>=3.7.1
numpy>=1.24.3
jinja2>=3.1.2
aiohttp>=3.8.4
async-timeout>=4.0.2
cryptography>=41.0.3
openai>=1.0.0
google-search-results>=2.4.0
dashscope>=1.0.0